Pull out Fields to global types
